+++ sos-webservices/sos-webservices-joc-global/src/main/java/com/sos/joc/bean/DailyPlanSummaryMBean.java
@@ -2,10 +2,10 @@ package com.sos.joc.bean;
 
 public interface DailyPlanSummaryMBean {
 
-    public float getfinished_orders();
-    public float getplanned_late_orders();
-    public float getplanned_orders();
-    public float getsubmitted_late_orders();
-    public float getsubmitted_orders();
+    public double getfinished_orders();
+    public double getplanned_late_orders();
+    public double getplanned_orders();
+    public double getsubmitted_late_orders();
+    public double getsubmitted_orders();
     
 }
